experiments: adds json parsing for use as custom menu config resolves #1
