+++ config.json
@@ -1,0 +1,56 @@
+{
+  "name": "DUI",
+  "type": "node",
+  "children": [
+    {
+      "name": "Scripts",
+      "type": "node",
+      "children": [
+        {
+          "name": "Simple Scripts",
+          "type": "node",
+          "children": [
+            {
+              "name": "Show Date",
+              "type": "executable",
+              "path": "D:/golang/dui/scripts/date.sh"
+            },
+            {
+              "name": "ip",
+              "type": "executable",
+              "path": "D:/golang/dui/scripts/ip_info.sh"
+            },
+            {
+              "name": "ping",
+              "type": "executable",
+              "path": "D:/golang/dui/scripts/ping.sh"
+            },
+            {
+              "name": "sys_info",
+              "type": "executable",
+              "path": "D:/golang/dui/scripts/sys_info.sh"
+            }
+          ]
+        },
+
+        {
+          "name": "Fun Scripts",
+          "type": "node",
+          "children": [
+            {
+              "name": "Countdown",
+              "type": "executable",
+              "path": "D:/golang/dui/scripts/countdown.sh"
+            },
+            {
+              "name": "matrix",
+              "type": "executable",
+              "path": "D:/golang/dui/scripts/matrix.sh"
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
+
